Handle requirement of provider URL in configuration function
This is done because Terraform does not handle multiple required attributes well. If only one is set in the provider configuration, the other one is prompted for, but Terraform then forgets the value for the one that is defined in the code. This confusing user-experience can be avoided by handling the requirement of one of the attributes manually.

The result of this issue was that only setting the "url" attribute and providing the "token" through a prompt, would result in an error about "url" not being set.